+++ config/splat.pspeu.dra.yaml
@@ -54,8 +54,6 @@ segments:
       - [0xBDE0, c, dra_psp/BDE0]
       - [0xC0B0, c, dra_psp/C0B0] # O4
       - [0xE6A8, c, dra_psp/E6A8]
-      - [0xF4A0, c, dra_psp/F4A0]
-      - [0xF4D0, c, dra_psp/F4D0]
       - [0x10340, c, dra_psp/10340]
       - [0x10990, c, dra_psp/10990]
       - [0x10BF0, c, dra_psp/10BF0]
