--- src/dra_psp/F4A0.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/F4A0", func_800F53A4);
